backend/api/requirements.txt
tensorflow-macos; platform_system == "Darwin" and platform_machine == "arm64" 
tensorflow; platform_system != "Darwin" or platform_machine != "arm64" 
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy
